Added some constraints to product location page

+++ src/ProductsLocationPage.java
@@ -169,6 +169,34 @@ public class ProductsLocationPage extends JDialog {
                             "Duplicate Product Location ID",
                             JOptionPane.ERROR_MESSAGE);
                 }
+                else if(Location_id.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Location ID field is empty, please enter a Locaiton ID",
+                            "Empty Location ID",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(Aisle.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Aisle field is empty, please enter a Aisle",
+                            "Empty Aisle",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(Shelf.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field is empty, please enter a Shelf",
+                            "Empty Shelf",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(!Shelf.matches("Top") || !Shelf.matches("Middle") || !Shelf.matches("Bottom"))
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field does not contain either Top, Middle, or Bottom is empty, please enter in either Top, Middle, or Bottom",
+                            "Wrong Shelf Data",
+                            JOptionPane.ERROR_MESSAGE);
+                }
                 else
                 {
                     try {
@@ -207,28 +235,56 @@ public class ProductsLocationPage extends JDialog {
                 String Aisle = productLocationAisleNumberTxtField.getText();
                 String Shelf = productLocationShelfTxtField.getText();
 
-                try {
-                    Connection connection = DriverManager.getConnection("jdbc:mysql://triton.towson.edu:3360/bdeguz1db", "bdeguz1", "COSC*bo29m");
-                    PreparedStatement preparedStatement = connection.prepareStatement("UPDATE LOCATION SET Aisle = ?, Shelf = ? WHERE Location_id = ?");
 
-                    preparedStatement.setString(1, Aisle);
-                    preparedStatement.setString(2, Shelf);
-                    preparedStatement.setString(3, Location_id);
-                    preparedStatement.executeUpdate();
-
-                    PreparedStatement selectStatement = connection.prepareStatement("SELECT * FROM LOCATION");
-                    ResultSet resultSet = selectStatement.executeQuery();
-                    productLocationTable.setModel(DbUtils.resultSetToTableModel(resultSet));
-
-                    productLocationIDTxtField.setText("");
-                    productLocationAisleNumberTxtField.setText("");
-                    productLocationShelfTxtField.setText("");
-                }
-                catch (SQLException e1)
+                if(Location_id.isEmpty())
                 {
-                    e1.printStackTrace();
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Location ID field is empty, please enter a Locaiton ID",
+                            "Empty Location ID",
+                            JOptionPane.ERROR_MESSAGE);
                 }
+                else if(Aisle.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Aisle field is empty, please enter a Aisle",
+                            "Empty Aisle",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(Shelf.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field is empty, please enter a Shelf",
+                            "Empty Shelf",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(!Shelf.matches("Top") || !Shelf.matches("Middle") || !Shelf.matches("Bottom"))
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field does not contain either Top, Middle, or Bottom is empty, please enter in either Top, Middle, or Bottom",
+                            "Wrong Shelf Data",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else {
+                    try {
+                        Connection connection = DriverManager.getConnection("jdbc:mysql://triton.towson.edu:3360/bdeguz1db", "bdeguz1", "COSC*bo29m");
+                        PreparedStatement preparedStatement = connection.prepareStatement("UPDATE LOCATION SET Aisle = ?, Shelf = ? WHERE Location_id = ?");
 
+                        preparedStatement.setString(1, Aisle);
+                        preparedStatement.setString(2, Shelf);
+                        preparedStatement.setString(3, Location_id);
+                        preparedStatement.executeUpdate();
+
+                        PreparedStatement selectStatement = connection.prepareStatement("SELECT * FROM LOCATION");
+                        ResultSet resultSet = selectStatement.executeQuery();
+                        productLocationTable.setModel(DbUtils.resultSetToTableModel(resultSet));
+
+                        productLocationIDTxtField.setText("");
+                        productLocationAisleNumberTxtField.setText("");
+                        productLocationShelfTxtField.setText("");
+                    } catch (SQLException e1) {
+                        e1.printStackTrace();
+                    }
+                }
             }
         });
 
@@ -240,26 +296,55 @@ public class ProductsLocationPage extends JDialog {
                 String Aisle = productLocationAisleNumberTxtField.getText();
                 String Shelf = productLocationShelfTxtField.getText();
 
-                try{
-                    Connection connection = DriverManager.getConnection("jdbc:mysql://triton.towson.edu:3360/bdeguz1db", "bdeguz1", "COSC*bo29m");
-                    PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM LOCATION WHERE Location_id = ? AND Aisle = ? AND Shelf = ?");
-
-                    preparedStatement.setString(1, Location_id);
-                    preparedStatement.setString(2, Aisle);
-                    preparedStatement.setString(3, Shelf);
-                    preparedStatement.executeUpdate();
-
-                    PreparedStatement selectStatement = connection.prepareStatement("SELECT * FROM LOCATION");
-                    ResultSet resultSet = selectStatement.executeQuery();
-                    productLocationTable.setModel(DbUtils.resultSetToTableModel(resultSet));
-
-                    productLocationIDTxtField.setText("");
-                    productLocationAisleNumberTxtField.setText("");
-                    productLocationShelfTxtField.setText("");
-                }
-                catch (SQLException e1)
+                if(Location_id.isEmpty())
                 {
-                    e1.printStackTrace();
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Location ID field is empty, please enter a Locaiton ID",
+                            "Empty Location ID",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(Aisle.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Aisle field is empty, please enter a Aisle",
+                            "Empty Aisle",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(Shelf.isEmpty())
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field is empty, please enter a Shelf",
+                            "Empty Shelf",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else if(!Shelf.matches("Top") || !Shelf.matches("Middle") || !Shelf.matches("Bottom"))
+                {
+                    JOptionPane.showMessageDialog(ProductsLocationPage.this,
+                            "Error: Shelf field does not contain either Top, Middle, or Bottom is empty, please enter in either Top, Middle, or Bottom",
+                            "Wrong Shelf Data",
+                            JOptionPane.ERROR_MESSAGE);
+                }
+                else {
+
+                    try {
+                        Connection connection = DriverManager.getConnection("jdbc:mysql://triton.towson.edu:3360/bdeguz1db", "bdeguz1", "COSC*bo29m");
+                        PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM LOCATION WHERE Location_id = ? AND Aisle = ? AND Shelf = ?");
+
+                        preparedStatement.setString(1, Location_id);
+                        preparedStatement.setString(2, Aisle);
+                        preparedStatement.setString(3, Shelf);
+                        preparedStatement.executeUpdate();
+
+                        PreparedStatement selectStatement = connection.prepareStatement("SELECT * FROM LOCATION");
+                        ResultSet resultSet = selectStatement.executeQuery();
+                        productLocationTable.setModel(DbUtils.resultSetToTableModel(resultSet));
+
+                        productLocationIDTxtField.setText("");
+                        productLocationAisleNumberTxtField.setText("");
+                        productLocationShelfTxtField.setText("");
+                    } catch (SQLException e1) {
+                        e1.printStackTrace();
+                    }
                 }
 
             }
